# api/openapi.yaml
openapi: 3.0.3
info:
  title: Fleetplane API
  description: >
    Agentless control plane for infrastructure fleets (fleetplane.io/v1alpha1).
    Every mutating request accepts an Idempotency-Key header; completed keys
    replay their stored response byte-identically.
  version: v1alpha1
servers:
  - url: /
security:
  - bearerAuth: []
paths:
  /v1/acquisitions:
    post:
      operationId: createAcquisition
      summary: Acquire capacity (find or create, 04 §4)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AcquireRequest" }
      responses:
        "201": { description: Acquisition accepted, content: { application/json: { schema: { $ref: "#/components/schemas/Acquisition" } } } }
        default: { $ref: "#/components/responses/Error" }
    get:
      operationId: listAcquisitions
      summary: List acquisitions (live ones only unless ?state says otherwise)
      parameters:
        - name: state
          in: query
          # Acquisitions are never garbage collected, so an unfiltered listing
          # would be the control plane's whole history: the default is the
          # live states (pending, provisioning, bound).
          description: "repeatable; default pending+provisioning+bound"
          style: form
          explode: true
          schema:
            type: array
            items: { type: string, enum: [pending, provisioning, bound, failed, released, expired] }
        - { name: resource, in: query, schema: { type: string } }
      responses:
        "200": { description: Acquisitions, content: { application/json: { schema: { $ref: "#/components/schemas/AcquisitionList" } } } }
        default: { $ref: "#/components/responses/Error" }
  /v1/acquisitions/{id}:
    get:
      operationId: getAcquisition
      summary: Read one acquisition
      parameters: [{ $ref: "#/components/parameters/id" }]
      responses:
        "200": { description: Acquisition, content: { application/json: { schema: { $ref: "#/components/schemas/Acquisition" } } } }
        default: { $ref: "#/components/responses/Error" }
    delete:
      operationId: releaseAcquisition
      summary: Release an acquisition (retry-safe)
      parameters: [{ $ref: "#/components/parameters/id" }]
      responses:
        "200": { description: Released }
        default: { $ref: "#/components/responses/Error" }
...
  /v1/pools:
    post:
      operationId: createPool
      summary: Create or update a pool by name
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PoolManifest" }
      responses:
        "200": { description: Pool, content: { application/json: { schema: { $ref: "#/components/schemas/Pool" } } } }
        default: { $ref: "#/components/responses/Error" }
    get:
      operationId: listPools
      summary: List pools
      responses:
        "200": { description: Pools, content: { application/json: { schema: { $ref: "#/components/schemas/PoolList" } } } }
        default: { $ref: "#/components/responses/Error" }
  /v1/pools/{id}:
    get:
      operationId: getPool
      summary: Read one pool
      parameters: [{ $ref: "#/components/parameters/id" }]
      responses:
        "200": { description: Pool, content: { application/json: { schema: { $ref: "#/components/schemas/Pool" } } } }
        default: { $ref: "#/components/responses/Error" }
    put:
      operationId: updatePool
      summary: Update a pool
      parameters: [{ $ref: "#/components/parameters/id" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PoolManifest" }
      responses:
        "200": { description: Pool, content: { application/json: { schema: { $ref: "#/components/schemas/Pool" } } } }
        default: { $ref: "#/components/responses/Error" }
    delete:
      operationId: deletePool
      summary: Delete a pool (409 while spec.replicas > 0 or any member resource still exists — scale to 0 and let the members drain first)
      parameters: [{ $ref: "#/components/parameters/id" }]
      responses:
        "200": { description: Deleted, content: { application/json: { schema: { type: object } } } }
        default: { $ref: "#/components/responses/Error" }
  /v1/pools/{id}:reconcile:
    post:
      operationId: reconcilePool
      summary: Trigger reconciliation (04 §3; 409 while the pool is paused)
      parameters: [{ $ref: "#/components/parameters/id" }]
      responses:
        "202": { description: Reconciling }
        default: { $ref: "#/components/responses/Error" }
  /v1/pools/{id}:pause:
    post:
      operationId: pausePool
      summary: Stop reconciling a pool (200 — nothing is left in flight; already paused is 200)
      parameters: [{ $ref: "#/components/parameters/id" }]
      responses:
        "200": { description: Paused, content: { application/json: { schema: { type: object } } } }
        default: { $ref: "#/components/responses/Error" }
  /v1/pools/{id}:resume:
    post:
      operationId: resumePool
      summary: Resume reconciling a pool and kick a reconcile (200; already running is 200)
      parameters: [{ $ref: "#/components/parameters/id" }]
      responses:
        "200": { description: Running, content: { application/json: { schema: { type: object } } } }
        default: { $ref: "#/components/responses/Error" }
  /v1/operations:
    get:
      operationId: listOperations
      summary: List open operations
      responses:
        "200": { description: Operations }
        default: { $ref: "#/components/responses/Error" }
  /v1/operations/{id}:
    get:
      operationId: getOperation
      summary: Read one operation
      parameters: [{ $ref: "#/components/parameters/id" }]
      responses:
        "200": { description: Operation }
        default: { $ref: "#/components/responses/Error" }
  /v1/operations/{id}:resolve:
    post:
      operationId: resolveOperation
      summary: Resolve an uncertain operation (operator attestation)
      parameters: [{ $ref: "#/components/parameters/id" }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [action]
              properties:
                action: { type: string, enum: [retry-verification, mark-failed] }
      responses:
        "202": { description: Resolution applied }
        default: { $ref: "#/components/responses/Error" }
  /v1/events:
    get:
      operationId: listEvents
      summary: List audit events (cursor = evt_ id)
      parameters:
        - { name: after, in: query, schema: { type: string } }
        - { name: since, in: query, schema: { type: string, example: "1h" } }
        - { name: limit, in: query, schema: { type: integer } }
        - { name: resource, in: query, schema: { type: string } }
      responses:
        "200": { description: Events }
        default: { $ref: "#/components/responses/Error" }
  /v1/providers:
    get:
      operationId: listProviders
      summary: Provider instance health (07 §8 — never part of readiness)
      responses:
        "200": { description: Providers }
        default: { $ref: "#/components/responses/Error" }
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, description: "flp_<id>.<secret> (ADR-008)" }
  parameters:
    id: { name: id, in: path, required: true, schema: { type: string } }
...
  responses:
    Error:
      description: Error envelope (single shape for every non-2xx)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorBody" }
  schemas:
    ErrorBody:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
            requestId: { type: string }
            retryable: { type: boolean }
            retryAfterSeconds: { type: integer }
            details: { type: object }
    Metadata:
      type: object
      required: [id]
      properties:
        id: { type: string }
        name: { type: string }
        labels: { type: object, additionalProperties: { type: string } }
        ownership: { type: string, enum: [managed, adopted, observed] }
        protected: { type: boolean }
        generation: { type: integer }
        observedGeneration: { type: integer }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        deletedAt: { type: string, format: date-time, nullable: true }
...
    AcquireRequest:
      type: object
      # The server rejects unknown fields; the spec says so too, so
      # spec-validated clients can't drift (docs/11 review finding).
      additionalProperties: false
      properties:
        kind: { type: string }
        class: { type: string }
        quantity: { type: integer, default: 1 }
        constraints:
          type: object
          additionalProperties:
            type: object
            properties: { min: { type: integer } }
        exclusive: { type: boolean }
        lease:
          type: object
          properties: { ttl: { type: string, example: "90m" } }
        # Queue budget (docs/11 §8): wait for existing capacity up to this
        # long before scaling up; empty = class default; "0s" = never queue.
        maxWait: { type: string, example: "10m" }
        idempotencyKey: { type: string }
    Acquisition:
      type: object
      required: [id, state]
      properties:
        apiVersion: { type: string }
        kind: { type: string }
        id: { type: string }
        state: { type: string, enum: [pending, provisioning, bound, failed, released, expired] }
        class: { type: string }
        resourceKind: { type: string }
        resourceId: { type: string }
        leaseId: { type: string }
        actor: { type: string }
        maxWait: { type: string, description: "resolved queue budget (docs/11 §8)" }
        queueDeadline: { type: string, format: date-time, description: "createdAt + maxWait; after this the acquisition scales up" }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    AcquisitionList:
      type: object
      properties:
        apiVersion: { type: string }
        kind: { type: string }
        items: { type: array, items: { $ref: "#/components/schemas/Acquisition" } }
...
    PoolManifest:
      type: object
      required: [spec]
      properties:
        apiVersion: { type: string }
        kind: { type: string }
        metadata: { $ref: "#/components/schemas/Metadata" }
        spec: { type: object, description: "reconcile.PoolSpec: class|kind+provider+machine, replicas, minReady, maxResources, reclaim.idleAfter" }
    Pool:
      type: object
      properties:
        apiVersion: { type: string }
        kind: { type: string }
        metadata: { $ref: "#/components/schemas/Metadata" }
        spec: { type: object }
        # Not part of PoolManifest: 'paused' has no "unset", so honouring it on
        # apply would let a manifest that omits the field silently resume a
        # paused pool.
        paused: { type: boolean, readOnly: true, description: "set via :pause / :resume; ignored on apply" }
    PoolList:
      type: object
      properties:
        apiVersion: { type: string }
        kind: { type: string }
        items: { type: array, items: { $ref: "#/components/schemas/Pool" } }